LayoutTests/imported/w3c/web-platform-tests/navigation-api/state/cross-document-away-and-back-expected.txt



FAIL entry.getState() behavior after navigating away and back assert_equals: expected 1 but got 0
